format StateSyncStatus in Debug using alternate flag
